requirements.txt
fastapi
uvicorn
requests
python-dotenv
pydantic
langchain
langchain-community
chromadb
huggingface-hub
sentence-transformers
